+++ bundles/org.eclipse.swt/Eclipse SWT PI/cocoa/org/eclipse/swt/internal/cocoa/NSCalendarDate.java
@@ -25,51 +25,51 @@ public NSCalendarDate(id id) {
 }
 
 public static NSCalendarDate calendarDate() {
-	int result = OS.objc_msgSend(OS.class_NSCalendarDate, OS.sel_calendarDate);
+	int /*long*/ result = OS.objc_msgSend(OS.class_NSCalendarDate, OS.sel_calendarDate);
 	return result != 0 ? new NSCalendarDate(result) : null;
 }
 
-public static NSCalendarDate dateWithYear(int year, int month, int day, int hour, int minute, int second, NSTimeZone aTimeZone) {
-	int result = OS.objc_msgSend(OS.class_NSCalendarDate, OS.sel_dateWithYear_month_day_hour_minute_second_timeZone_, year, month, day, hour, minute, second, aTimeZone != null ? aTimeZone.id : 0);
+public static NSCalendarDate dateWithYear(int /*long*/ year, int /*long*/ month, int /*long*/ day, int /*long*/ hour, int /*long*/ minute, int /*long*/ second, NSTimeZone aTimeZone) {
+	int /*long*/ result = OS.objc_msgSend(OS.class_NSCalendarDate, OS.sel_dateWithYear_month_day_hour_minute_second_timeZone_, year, month, day, hour, minute, second, aTimeZone != null ? aTimeZone.id : 0);
 	return result != 0 ? new NSCalendarDate(result) : null;
 }
 
-public int dayOfMonth() {
-	return OS.objc_msgSend(this.id, OS.sel_dayOfMonth);
+public int /*long*/ dayOfMonth() {
+	return (int)/*64*/OS.objc_msgSend(this.id, OS.sel_dayOfMonth);
 }
 
-public int hourOfDay() {
-	return OS.objc_msgSend(this.id, OS.sel_hourOfDay);
+public int /*long*/ hourOfDay() {
+	return (int)/*64*/OS.objc_msgSend(this.id, OS.sel_hourOfDay);
 }
 
-public int minuteOfHour() {
-	return OS.objc_msgSend(this.id, OS.sel_minuteOfHour);
+public int /*long*/ minuteOfHour() {
+	return (int)/*64*/OS.objc_msgSend(this.id, OS.sel_minuteOfHour);
 }
 
-public int monthOfYear() {
-	return OS.objc_msgSend(this.id, OS.sel_monthOfYear);
+public int /*long*/ monthOfYear() {
+	return (int)/*64*/OS.objc_msgSend(this.id, OS.sel_monthOfYear);
 }
 
-public int secondOfMinute() {
-	return OS.objc_msgSend(this.id, OS.sel_secondOfMinute);
+public int /*long*/ secondOfMinute() {
+	return (int)/*64*/OS.objc_msgSend(this.id, OS.sel_secondOfMinute);
 }
 
 public NSTimeZone timeZone() {
-	int result = OS.objc_msgSend(this.id, OS.sel_timeZone);
+	int /*long*/ result = OS.objc_msgSend(this.id, OS.sel_timeZone);
 	return result != 0 ? new NSTimeZone(result) : null;
 }
 
-public int yearOfCommonEra() {
-	return OS.objc_msgSend(this.id, OS.sel_yearOfCommonEra);
+public int /*long*/ yearOfCommonEra() {
+	return (int)/*64*/OS.objc_msgSend(this.id, OS.sel_yearOfCommonEra);
 }
 
 public static NSDate dateWithTimeIntervalSinceNow(double secs) {
-	int result = OS.objc_msgSend(OS.class_NSCalendarDate, OS.sel_dateWithTimeIntervalSinceNow_, secs);
+	int /*long*/ result = OS.objc_msgSend(OS.class_NSCalendarDate, OS.sel_dateWithTimeIntervalSinceNow_, secs);
 	return result != 0 ? new NSCalendarDate(result) : null;
 }
 
 public static NSDate distantFuture() {
-	int result = OS.objc_msgSend(OS.class_NSCalendarDate, OS.sel_distantFuture);
+	int /*long*/ result = OS.objc_msgSend(OS.class_NSCalendarDate, OS.sel_distantFuture);
 	return result != 0 ? new NSCalendarDate(result) : null;
 }
